http 형식 수정, -> viewmodel을 통해 response 를 json형식으로 접근함
