+++ app/src/main/java/com/example/thanksd/httpconnection/JsonViewModel.kt
@@ -1,0 +1,9 @@
+package com.example.thanksd.httpconnection
+
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import org.json.JSONObject
+
+class JsonViewModel:ViewModel() {
+    val response = MutableLiveData<JSONObject>()
+}
