+++ app/src/main/java/com/example/thanksd/login/google/GoogleLoginActivity.kt
@@ -1,5 +1,6 @@
 package com.example.thanksd.login.google
 
+import android.app.Activity
 import android.content.Intent
 import android.net.Uri
 import android.os.Build
@@ -19,11 +20,16 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.viewinterop.AndroidView
+import androidx.lifecycle.Observer
+import androidx.lifecycle.findViewTreeLifecycleOwner
 import com.example.thanksd.BuildConfig
 import com.example.thanksd.MainPage.MainActivity
 import com.example.thanksd.httpconnection.HttpFunc
+import com.example.thanksd.httpconnection.JsonViewModel
 import com.example.thanksd.login.LoginActivity
+import com.example.thanksd.login.dataclass.ClientInformation
 import com.example.thanksd.login.google.ui.theme.ThanksDTheme
 
 
@@ -50,6 +56,7 @@ class GoogleLoginActivity : ComponentActivity() {
 fun GoogleLoginScreen() {
     // 웹뷰를 표시
     val cur = LocalContext.current
+    val lifecycleOwner = LocalLifecycleOwner.current
     val redirecturl = BuildConfig.Redirecturl
     val clientId = BuildConfig.Client_id
     var check = remember {
@@ -58,9 +65,9 @@ fun GoogleLoginScreen() {
     val url = "https://accounts.google.com/o/oauth2/v2/auth?response_type=code&client_id="+
             clientId+"&redirect_uri="+redirecturl+"&scope=email profile"
 
-    val cookieManager = CookieManager.getInstance()
-    cookieManager.removeAllCookies(null)
-    cookieManager.flush()
+//    val cookieManager = CookieManager.getInstance()
+//    cookieManager.removeAllCookies(null)
+//    cookieManager.flush()
     LaunchedEffect(check.value){
         if(check.value == 1){
             cur.startActivity(Intent(cur, MainActivity::class.java))
@@ -68,7 +75,7 @@ fun GoogleLoginScreen() {
             cur.startActivity(Intent(cur, LoginActivity::class.java))
         }
     }
-
+    val response = JsonViewModel()
     AndroidView(
         modifier = Modifier.fillMaxSize(),
         factory = { context ->
@@ -85,9 +92,8 @@ fun GoogleLoginScreen() {
                                 val JsonObj = JSONObject()
                                 val httpManager = HttpFunc(serverurl)
                                 JsonObj.put("code",authCode)
-                                httpManager.POST(JsonObj)
+                                httpManager.POST(JsonObj,response)
 //                                accessToken.value = authCode
-                                check.value = 1
                                 // 사용자 정보 저장 코드 필요
                                 return true
                             }
@@ -101,6 +107,7 @@ fun GoogleLoginScreen() {
                     }
 
                 }
+
                 val userAgentString = this.settings.userAgentString
                 val newUserAgentString = userAgentString
                     .replace("; wv", "")
@@ -115,8 +122,19 @@ fun GoogleLoginScreen() {
 //                Log.d("google_login", "check")
             }
 
+
         }
     )
+    response.response.observe(lifecycleOwner, Observer {response->
+        val data = JSONObject(response?.get("data").toString())
+        val token = data.get("token").toString()
+        val email = data.get("email").toString()
+        val isRegistered = data.get("isRegistered").toString().toBoolean()
+        val platformId = data.get("platformId").toString()
+        ClientInformation.updateValue(token,email,isRegistered,platformId)
+        check.value = 1
+        Log.d("check12", ClientInformation.platformID)
+    })
 }
